changes to submodules, ci
