+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "public/uv"]
 	path = public/uv
-	url = git@github.com:inertia-unblocker/uv-scripts
+	url = https://github.com/inertia-unblocker/uv-scripts
